requirements.txt
asgiref==3.5.0
attrs==21.4.0
certifi==2021.10.8
charset-normalizer==2.0.12
Django==4.0.3
django-filter==21.1
djangorestframework==3.13.1
drf-spectacular==0.21.2
idna==3.3
inflection==0.5.1
jsonschema==4.4.0
pyrsistent==0.18.1
python-dotenv==0.19.2
pytz==2021.3
PyYAML==6.0
requests==2.27.1
sqlparse==0.4.2
uritemplate==4.1.1
urllib3==1.26.8
